Moved each renderer and corresponding tests into separate files.
