+++ result_redirect.go
@@ -1,0 +1,12 @@
+package detour
+
+import "net/http"
+
+type RedirectResult struct {
+	Location   string
+	StatusCode int
+}
+
+func (this RedirectResult) Render(response http.ResponseWriter, request *http.Request) {
+	http.Redirect(response, request, this.Location, this.StatusCode)
+}
