+++ result_status_code.go
@@ -1,0 +1,13 @@
+package detour
+
+import "net/http"
+
+type StatusCodeResult struct {
+	StatusCode int
+	Message    string
+}
+
+func (this StatusCodeResult) Render(response http.ResponseWriter, request *http.Request) {
+	writeContentTypeAndStatusCode(response, this.StatusCode, plaintextContentType)
+	response.Write([]byte(this.Message))
+}
